feat: addLayoutWrap-andHeaderToAllPages-changeRoutingFile-refactorMainPageCode
