+++ src/components/Layout/Layout.scss
@@ -1,0 +1,5 @@
+.layout__container {
+  max-width: 1440px;
+  width: 100%;
+  margin: 0 auto;
+}
