+++ src/components/_Header/_Header.scss
@@ -1,0 +1,73 @@
+@mixin DisplayFlex {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.header {
+  @include DisplayFlex;
+  max-width: 100%;
+  width: 100%;
+  height: 77px;
+  border-bottom: 1px solid rgba(70, 163, 88, 0.2);
+
+  .header_logo {
+    max-width: 150px;
+    width: 100%;
+  }
+
+  .header__nav {
+    @include DisplayFlex;
+    flex-grow: 0.25;
+
+    .header__nav_link-home,
+    .header__nav_link-catalog,
+    .header__nav_link-about {
+      font-size: 18px;
+      font-weight: 400;
+      line-height: 20px;
+      color: rgb(61, 61, 61);
+      text-decoration: none;
+      outline: none;
+      transition: all 0.5s ease;
+
+      &:hover {
+        color: rgb(70, 163, 88);
+      }
+
+      &.active {
+        position: relative;
+        font-size: 18px;
+        font-weight: 700;
+        line-height: 20px;
+
+        &::before {
+          content: "";
+          position: absolute;
+          left: 0;
+          bottom: -29px;
+          width: 100%;
+          height: 3px;
+          background-color: rgb(70, 163, 88);
+        }
+      }
+    }
+  }
+
+  .header__inner {
+    @include DisplayFlex;
+    flex-grow: 0.05;
+
+    .header__inner_link-login,
+    .header__inner_link-register,
+    .header__inner_link-logout,
+    .header__inner_link-basket {
+      height: 40px;
+
+      img {
+        max-width: 40px;
+        width: 100%;
+      }
+    }
+  }
+}
