+++ src/pages/AboutPage/AboutPage.scss
@@ -1,4 +1,4 @@
-.main-page {
+.about-page {
   display: flex;
   flex-direction: column;
   align-items: center;
