+++ src/pages/AboutPage/AboutPage.tsx
@@ -1,0 +1,12 @@
+import "./AboutPage.scss";
+import React, { ReactElement } from "react";
+
+function AboutPage(): ReactElement {
+  return (
+    <div className="about-page" style={{ marginTop: "50px" }}>
+      AboutPage Comming Soon...
+    </div>
+  );
+}
+
+export default AboutPage;
