+++ src/pages/CatalogPage/Catalog.scss
@@ -1,4 +1,4 @@
-.main-page {
+.catalog-page {
   display: flex;
   flex-direction: column;
   align-items: center;
