+++ src/pages/CatalogPage/Catalog.tsx
@@ -1,0 +1,12 @@
+import "./Catalog.scss";
+import React, { ReactElement } from "react";
+
+function CatalogPage(): ReactElement {
+  return (
+    <div className="catalog-page" style={{ marginTop: "50px" }}>
+      CatalogPage Comming Soon...
+    </div>
+  );
+}
+
+export default CatalogPage;
